[refactor]: apply lints
