+++ client/benches/torii.rs
@@ -22,7 +22,7 @@ fn query_requests(criterion: &mut Criterion) {
 
     let chain_id = get_chain_id();
     let configuration = get_config(
-        unique_vec![peer.id.clone()],
+        &unique_vec![peer.id.clone()],
         Some(chain_id.clone()),
         Some(get_key_pair()),
     );
@@ -132,7 +132,7 @@ fn instruction_submits(criterion: &mut Criterion) {
 
     let chain_id = get_chain_id();
     let configuration = get_config(
-        unique_vec![peer.id.clone()],
+        &unique_vec![peer.id.clone()],
         Some(chain_id.clone()),
         Some(get_key_pair()),
     );
